#  add AES-GCM-SIV amd64 assembler code for AES-CTR and POLYVAL 